splitting data query types
